LayoutTests/fullscreen/full-screen-iframe-without-allow-attribute-allowed-from-parent-expected.txt

CONSOLE MESSAGE: line 1: Feature policy 'Fullscreen' check failed for iframe with origin 'null' and allow attribute ''.
Test entering full screen security restrictions. An iframe without an allow attribute is still permitted to fullscreen if the request comes from the containing document.

To test manually, press any key - the page should enter full screen mode.

TEST(document.webkitFullscreenEnabled) OK
iframe's webkitFullscreenEnabled should be false:
TEST(iframeMessage == 'document.webkitFullscreenEnabled == false') OK
EVENT(webkitfullscreenchange)
SUCCEED - entered full screen!
END OF TEST
